Add terraform documentation, makefile and restructuration of tests files
