+++ Makefile
@@ -1,0 +1,14 @@
+testacc:
+	TF_ACC=1 go test ./...
+
+.PHONY: provider_code_spec.json
+provider_code_spec.json:
+	tfplugingen-openapi generate --config ./generator_config.yml --output ./provider_code_spec.json ./openapi/logto-openapi-source.json
+
+.PHONY: provider
+provider:
+	tfplugingen-framework generate all --input ./provider_code_spec.json --output internal/provider
+
+.PHONY: docs
+docs:
+	tfplugindocs generate
