+++ docs/index.md
@@ -1,0 +1,28 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "logto Provider"
+subcategory: ""
+description: |-
+  Logto terraform provider developed by Lenstra.
+---
+
+# logto Provider
+
+Logto terraform provider developed by Lenstra.
+
+## Example Usage
+
+```terraform
+provider "logto" {
+  tenant_id    = "yourTenantId"
+  access_token = "yourAccessToken"
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Optional
+
+- `access_token` (String) API key for you instance, can be set as environment variable LOGTO_ACCESS_TOKEN
+- `tenant_id` (String) API tenant_id for you instance, can be set as environment variable LOGTO_TENANT_ID
