+++ docs/resources/application.md
@@ -1,0 +1,38 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "logto_application Resource - terraform-provider-logto"
+subcategory: ""
+description: |-
+  
+---
+
+# logto_application (Resource)
+
+
+
+## Example Usage
+
+```terraform
+resource "logto_application" "app" {
+  name        = "test"
+  description = "test app description"
+  type        = "Native"
+}
+```
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `name` (String)
+- `type` (String)
+
+### Optional
+
+- `description` (String)
+
+### Read-Only
+
+- `id` (String) The unique identifier of the application.
+- `tenant_id` (String)
