+++ examples/provider/provider.tf
@@ -1,0 +1,5 @@
+
+provider "logto" {
+  tenant_id    = "yourTenantId"
+  access_token = "yourAccessToken"
+}
