+++ examples/resources/logto_application/resource.tf
@@ -1,0 +1,6 @@
+
+resource "logto_application" "app" {
+  name        = "test"
+  description = "test app description"
+  type        = "Native"
+}
