+++ tools/tools.go
@@ -1,0 +1,8 @@
+//go:build tools
+
+package tools
+
+import (
+	// Documentation generation
+	"github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs"
+)
